fixup! [#250] restriction of personal settings for organiser and event with only us as attnedees
