+++ src/features/Events/EventUpdateModal.tsx
@@ -1,21 +1,13 @@
-import React, { useCallback, useEffect, useRef, useState } from 'react'
 import { useAppSelector } from '@/app/hooks'
 import { ResponsiveDialog } from '@/components/Dialog'
 import EventFormFields from '@/components/Event/EventFormFields'
-import type { EventFormHandle } from '@/components/Event/EventFormFields.types'
-import {
-  clearEventFormTempData,
-  EventFormContext
-} from '@/utils/eventFormTempStorage'
-import { useI18n } from 'twake-i18n'
-import { CalendarEvent } from './EventsTypes'
 import { useScreenSizeDetection } from '@/useScreenSizeDetection'
-import { EventActions } from './EventActions'
-import { useBuildInitialValues } from '@/components/Event/hooks/useBuildInitialValues'
-import { useSubmitUpdateEvent } from '@/features/Events/hooks/useSubmitUpdateEvent'
-import { useMasterEvent } from '@/features/Events/hooks/useMasterEvent'
-import { useUserPersonalCalendars } from '@/features/Calendars/hooks/useUserPersonalCalendars'
 import type { SxProps } from '@mui/material/styles'
+import React from 'react'
+import { useI18n } from 'twake-i18n'
+import { EventActions } from './EventActions'
+import { CalendarEvent } from './EventsTypes'
+import { useEventUpdateModal } from './useEventUpdateModal'
 
 const dialogPaddingStyles = (isMobile: boolean): SxProps => ({
   '& .MuiDialogActions-root': {
@@ -24,7 +16,7 @@ const dialogPaddingStyles = (isMobile: boolean): SxProps => ({
   }
 })
 
-interface EventUpdateModalProps {
+export interface EventUpdateModalProps {
   eventId: string
   calId: string
   open: boolean
@@ -37,79 +29,31 @@ interface EventUpdateModalProps {
 
 const EventUpdateModalInternal: React.FC<
   EventUpdateModalProps & { event: CalendarEvent }
-> = ({
-  eventId,
-  calId,
-  open,
-  onClose,
-  isSpecific,
-  onCloseAll,
-  event,
-  typeOfAction
-}) => {
+> = props => {
+  const { open, isSpecific, event, typeOfAction } = props
   const { t } = useI18n()
   const { isTooSmall: isMobile } = useScreenSizeDetection()
-  const calList = useAppSelector(state => state.calendars.list)
-  const user = useAppSelector(state => state.user)
 
-  const userPersonalCalendars = useUserPersonalCalendars(
-    calList,
-    user.userData?.openpaasId
-  )
-
-  const [showMore, setShowMore] = useState(false)
-  const formRef = useRef<EventFormHandle>(null)
-
-  const { masterEvent, effectiveEvent } = useMasterEvent(
-    event,
-    open,
-    typeOfAction
-  )
-
-  const initialValues = useBuildInitialValues({
-    event: effectiveEvent || null,
-    calId,
-    selectedRange: null
-  })
-
-  useEffect(() => {
-    const resetShowMore = (): void => {
-      if (!open) setShowMore(false)
-    }
-    resetShowMore()
-  }, [open])
-
-  const handleClose = useCallback((): void => {
-    clearEventFormTempData('update')
-    if (onCloseAll) onCloseAll()
-    else onClose({}, 'backdropClick')
-    setShowMore(false)
-  }, [onClose, onCloseAll])
-
-  const { handleSubmit } = useSubmitUpdateEvent({
-    event,
-    calId,
-    eventId,
-    typeOfAction,
-    masterEvent,
+  const {
+    userPersonalCalendars,
     showMore,
-    onClose: handleClose
-  })
-
-  const tempContext: EventFormContext = { eventId, calId, typeOfAction }
-
-  const handleExpandToggle = !isSpecific
-    ? (): void => setShowMore(s => !s)
-    : undefined
+    setShowMore,
+    formRef,
+    effectiveEvent,
+    initialValues,
+    handleClose,
+    handleSubmit,
+    handleExpandToggle,
+    handleSave,
+    tempContext
+  } = useEventUpdateModal(props)
 
   const actions = (
     <EventActions
       showExpandedBtn={!showMore && !isSpecific}
       isEdit
       onClose={handleClose}
-      onSave={async () => {
-        await formRef.current?.submit()
-      }}
+      onSave={handleSave}
       onExpanded={() => setShowMore(s => !s)}
     />
   )
